NCI-Agency/anet#3342: Move notes to AbstractCustomizableAnetBean
Reduce the number of classes that have notes to those that need them.
