+++ src/main/java/mil/dds/anet/beans/NoteRelatedObject.java
@@ -6,7 +6,6 @@ import io.leangen.graphql.annotations.GraphQLInputField;
 import io.leangen.graphql.annotations.GraphQLQuery;
 import io.leangen.graphql.annotations.GraphQLRootContext;
 import java.time.Instant;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import javax.ws.rs.WebApplicationException;
@@ -61,25 +60,6 @@ public class NoteRelatedObject extends AbstractAnetBean {
 
   @Override
   public void setUpdatedAt(Instant updatedAt) {
-    // just ignore
-  }
-
-  @Override
-  public CompletableFuture<List<Note>> loadNotes(@GraphQLRootContext Map<String, Object> context) {
-    throw new WebApplicationException("no notes field on NoteRelatedObject");
-  }
-
-  @Override
-  @JsonIgnore
-  @GraphQLIgnore
-  public List<Note> getNotes() {
-    throw new WebApplicationException("no notes field on NoteRelatedObject");
-  }
-
-  @Override
-  @JsonIgnore
-  @GraphQLIgnore
-  public void setNotes(List<Note> notes) {
     // just ignore
   }
