+++ src/main/java/mil/dds/anet/views/AbstractAnetBean.java
@@ -3,13 +3,7 @@ package mil.dds.anet.views;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.leangen.graphql.annotations.GraphQLInputField;
 import io.leangen.graphql.annotations.GraphQLQuery;
-import io.leangen.graphql.annotations.GraphQLRootContext;
 import java.time.Instant;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import mil.dds.anet.AnetObjectEngine;
-import mil.dds.anet.beans.Note;
 
 public abstract class AbstractAnetBean {
 
@@ -22,7 +16,7 @@ public abstract class AbstractAnetBean {
   @GraphQLQuery
   @GraphQLInputField
   protected Instant updatedAt;
-  private List<Note> notes;
+  // annotated below
   private String batchUuid;
 
   public AbstractAnetBean() {
@@ -51,27 +45,6 @@ public abstract class AbstractAnetBean {
 
   public void setUpdatedAt(Instant updatedAt) {
     this.updatedAt = updatedAt;
-  }
-
-  @GraphQLQuery(name = "notes")
-  public CompletableFuture<List<Note>> loadNotes(@GraphQLRootContext Map<String, Object> context) {
-    if (notes != null) {
-      return CompletableFuture.completedFuture(notes);
-    }
-    return AnetObjectEngine.getInstance().getNoteDao().getNotesForRelatedObject(context, uuid)
-        .thenApply(o -> {
-          notes = o;
-          return o;
-        });
-  }
-
-  public List<Note> getNotes() {
-    return notes;
-  }
-
-  @GraphQLInputField(name = "notes")
-  public void setNotes(List<Note> notes) {
-    this.notes = notes;
   }
 
   @JsonIgnore
